1. 테스트 케이스 추가

+++ server/src/main/java/com/anamensis/server/entity/File.java
@@ -12,7 +12,7 @@ import java.time.LocalDateTime;
 public class File {
     private long id;
 
-    private int tableCodePk;
+    private long tableCodePk;
 
     private long tableRefPk;
